pyproject.toml
[tool.poetry]
name = "openinsider-scraper"
version = "0.1.0"
description = "High-performance web scraper for OpenInsider.com"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
lxml = "^5.1.0"  # Parser for BeautifulSoup
PyYAML = "^6.0.1"
ratelimit = "^2.2.1"
tenacity = "^8.2.3" # For more flexible retry logic than requests' built-in
pybreaker = "^1.0.0"
selenium = "^4.15.0" # Optional, for Selenium fallback
pydantic = "^2.5.3" # For data validation and models
python-dateutil = "^2.8.2" # For flexible date parsing

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.12.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
openinsider-cli = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true # For some 3rd party libs if stubs are missing
# strict = true # Uncomment for maximal strictness